requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn==0.38.0

# HTTP Client
httpx==0.25.1

# Database (PostgreSQL/Supabase)
sqlalchemy>=2.0.36
asyncpg>=0.29.0
aiosqlite>=0.19.0 # Added for SQLite async support

# Data Validation
pydantic>=2.5.0
email-validator==2.3.0

# Templating
jinja2==3.1.2

# File Uploads
python-multipart==0.0.20

# Timezone Support
pytz==2023.3

# Environment
python-dotenv

# JWT Authentication
pyjwt==2.10.1

# FastAPI Users for Authentication
fastapi-users[sqlalchemy]==15.0.1
